Cleanup: move api package up a level
